+++ pytmbot/handlers/server_handlers/cpu.py
@@ -27,6 +27,7 @@ from pytmbot.handlers.handlers_util.utils import (
 from pytmbot.handlers.server_handlers.inline.common import (
     build_user_bound_callback_data,
 )
+from pytmbot.keyboards.keyboards import NAV_SERVER
 from pytmbot.logs import Logger
 from pytmbot.parsers.compiler import Compiler
 
@@ -130,6 +131,7 @@ def handle_cpu(message: Message, bot: TeleBot) -> None:
             text=cpu_message,
             parse_mode="HTML",
             reply_markup=keyboard,
+            nav_keyboard=NAV_SERVER,
         )
         return None