+++ pytmbot/handlers/server_handlers/filesystem.py
@@ -25,6 +25,7 @@ from pytmbot.handlers.handlers_util.utils import (
 from pytmbot.handlers.server_handlers.inline.common import (
     build_user_bound_callback_data,
 )
+from pytmbot.keyboards.keyboards import NAV_SERVER
 from pytmbot.logs import Logger
 from pytmbot.parsers.compiler import Compiler
 
@@ -86,6 +87,7 @@ def handle_file_system(message: Message, bot: TeleBot) -> None:
             text=bot_answer,
             parse_mode="HTML",
             reply_markup=keyboard,
+            nav_keyboard=NAV_SERVER,
         )
         return None