+++ pytmbot/handlers/server_handlers/memory.py
@@ -21,6 +21,7 @@ from pytmbot.handlers.handlers_util.utils import (
     send_bot_message,
     send_server_message,
 )
+from pytmbot.keyboards.keyboards import NAV_SERVER
 from pytmbot.logs import Logger
 from pytmbot.parsers.compiler import Compiler
 
@@ -76,6 +77,7 @@ def handle_memory(message: Message, bot: TeleBot) -> None:
             text=bot_answer,
             reply_markup=keyboard,
             parse_mode="HTML",
+            nav_keyboard=NAV_SERVER,
         )
         return None