+++ pytmbot/handlers/server_handlers/sensors.py
@@ -26,6 +26,7 @@ from pytmbot.handlers.handlers_util.utils import (
 from pytmbot.handlers.server_handlers.inline.common import (
     build_user_bound_callback_data,
 )
+from pytmbot.keyboards.keyboards import NAV_SERVER
 from pytmbot.logs import Logger
 from pytmbot.parsers.compiler import Compiler
 
@@ -104,6 +105,7 @@ def handle_sensors(message: Message, bot: TeleBot) -> None:
             text=sensors_message,
             parse_mode="HTML",
             reply_markup=keyboard,
+            nav_keyboard=NAV_SERVER,
         )
 
     except Exception as error: