+++ pytmbot/handlers/server_handlers/uptime.py
@@ -24,6 +24,7 @@ from pytmbot.handlers.handlers_util.utils import (
 from pytmbot.handlers.server_handlers.inline.common import (
     build_user_bound_callback_data,
 )
+from pytmbot.keyboards.keyboards import NAV_SERVER
 from pytmbot.logs import Logger
 from pytmbot.parsers.compiler import Compiler
 
@@ -86,6 +87,7 @@ def handle_uptime(message: Message, bot: TeleBot) -> None:
             message.chat.id,
             text=bot_answer,
             reply_markup=keyboard,
+            nav_keyboard=NAV_SERVER,
         )
         return None